requirements.txt
psycopg2-binary>=2.9.0
jinja2>=3.0.0
click>=8.0.0
pyyaml>=6.0
